# Conflicts:
#	menu/menu_setting.c

+++ gfx/drivers/vulkan.c
@@ -3182,7 +3182,7 @@ static void vulkan_set_texture_frame(void *data,
       const void *frame, bool rgb32, unsigned width, unsigned height,
       float alpha)
 {
-   int y;
+   size_t y;
    unsigned stride;
    uint8_t *ptr                        = NULL;
    uint8_t *dst                        = NULL;
@@ -3244,7 +3244,7 @@ static void vulkan_set_texture_frame(void *data,
    {
       for (y = 0; y < height; y++, dst += texture->stride, src += stride)
       {
-         int x;
+         size_t x;
          uint16_t *srcpix = (uint16_t*)src;
          uint32_t *dstpix = (uint32_t*)dst;
          for (x = 0; x < width; x++, srcpix++, dstpix++)
@@ -3415,13 +3415,13 @@ static const video_poke_interface_t vulkan_poke_interface = {
    vulkan_load_texture,
    vulkan_unload_texture,
    vulkan_set_video_mode,
-   vulkan_get_refresh_rate, /* get_refresh_rate */
-   NULL,
+   vulkan_get_refresh_rate,            /* get_refresh_rate */
+   NULL,                               /* set_filtering */
    vulkan_get_video_output_size,
    vulkan_get_video_output_prev,
    vulkan_get_video_output_next,
-   NULL,
-   NULL,
+   NULL,                               /* get_current_framebuffer */
+   NULL,                               /* get_proc_address */
    vulkan_set_aspect_ratio,
    vulkan_apply_state_changes,
    vulkan_set_texture_frame,
